Split single file into multiple files
Split out `redisenterprise.go` into separate files to avoid having a single monolithic file
