+++ internal/plugin/new_user_test.go
